+++ src/webserver/folderart.c
@@ -29,9 +29,8 @@
  * @param mg_user_data pointer to mongoose configuration
  * @return true if an image was found, else false
  */
-bool request_handler_folderart(struct mg_connection *nc, struct mg_http_message *hm,
-        struct t_mg_user_data *mg_user_data)
-{
+bool request_handler_folderart(struct mg_connection *nc, struct mg_http_message *hm) {
+    struct t_mg_user_data *mg_user_data = (struct t_mg_user_data *) nc->mgr->userdata;
     if (sdslen(mg_user_data->music_directory) == 0) {
         webserver_redirect_placeholder_image(nc, PLACEHOLDER_NA);
         return false;