+++ src/webserver/tagart.c
@@ -30,9 +30,8 @@
  * @param mg_user_data webserver configuration
  * @return true on success, else false
  */
-bool request_handler_tagart(struct mg_connection *nc, struct mg_http_message *hm,
-        struct t_mg_user_data *mg_user_data)
-{
+bool request_handler_tagart(struct mg_connection *nc, struct mg_http_message *hm) {
+    struct t_mg_user_data *mg_user_data = (struct t_mg_user_data *) nc->mgr->userdata;
     struct t_config *config = mg_user_data->config;
     sds tag = get_uri_param(&hm->query, "tag=");
     sds value = get_uri_param(&hm->query, "value=");